Remove ProviderBatching from provider_model.go in the google package (#7850)
